Added some basic algebra "magic-methods"
The AstroImage class can now be passed to np.rad2deg, np.exp, np.log,
etc... and proper error-propagation will be performed.

Also added a method called "get_ticks" to estimate the proper tick
spacing, tick formatting, and minor-tick-frequency for the image at
hand. This is called by the show() method to make sure the shown images
have nice, clean, ticks. Finally, the default color-map was updated to
be "viridis," which is perceptually uniform, and thus superior to the
default "rainbow" colormap.