Minor updates
Added some exceptions for when there might be a problem scaling the
uncertainty array in the "scale" method using the (unconventional)
"SBSCALE" header keyword. But I mean, come on, we all know that scaling
constants MUST have some uncertainty, so we should probably account for
those.

Also added instructions on how to INSTALL the package directly from
github.