Added scaling tracker attribute
There is now a "protected" atribute "_scaling_quantity," which trackes
whether the images has been scaled to flux or intensity units. If it HAS
been scaled, then invoking the  img.scale() method again will return the
array to its unscaled values and chang the "_scaling_quantity" attribute
to a None value to indicate that the current contents of img.arr are not
scaled by the BSCALE and BZERO keywords.