Complete re-write of the AstroImage package
The package is DRAMATICALLY more modularized.

There is still much that could be done to modularize the larger classes
(BaseImage and AstroImage).

The name of the "AstroImage" class may be soon changed to
"ReducedScience" to avoid confusion between the package name and the
class names.

The image alignment and combining code has been moved into the
"ImageStack" class.

The Astrometry.net functionality has been moved into a separet
"AstrometrySolver" class.