Adjusted 'BZERO' issue to match new Astropy version
Older versions of Astropy treated the 'BZERO' keyword in FITS headers a
little differently.

I now allow the BaseImage.read method to scale up array values to match
the expected values. Then I set the 'BZERO' keyword to zero so that when
the data is saved to disk, it is not somehow changed...